Fix PHP 8.4 FrameworkBundle grind: DI merge serialization, strict types, and harness fixes.
Handle closure factories during extension config merge, cast Parameter and
SplFileInfo usages under strict typing, and align factory definitions and
shared component coercions required for FrameworkBundle functional tests.

+++ src/Symfony/Component/Process/Pipes/WindowsPipes.php
@@ -141,7 +141,7 @@ class WindowsPipes extends AbstractPipes
 
         if ($blocking) {
             if ($w) {
-                @stream_select($r, $w, $e, 0, Process::TIMEOUT_PRECISION * 1E6);
+                @stream_select($r, $w, $e, 0, (int) (Process::TIMEOUT_PRECISION * 1E6));
             } elseif ($this->fileHandles) {
                 usleep((int) (Process::TIMEOUT_PRECISION * 1E6));
             }
